change data type from `uint64_t` -> `size_t` (#30)
and
-    always use `hwthread` id's numbering in debug statements

+++ include/allocator.h
@@ -6,12 +6,12 @@
 
 #include "test_types.h"
 
-static inline uint64_t get_cl_size()
+static inline size_t get_cl_size()
 {
-    static uint64_t c_cl_size = 0;
+    static size_t c_cl_size = 0;
     if (c_cl_size == 0)
     {
-        uint64_t cl_size = (uint64_t)sysconf(_SC_LEVEL1_DCACHE_LINESIZE);
+        size_t cl_size = (size_t)sysconf(_SC_LEVEL1_DCACHE_LINESIZE);
         // fprintf(stdout, "CL_SIZE from SYSCONF: %zu Bytes\n", cl_size);
         c_cl_size = ((cl_size > 0 && !(cl_size & 7)) ? cl_size : 64);
         // fprintf(stdout, "Final CL_SIZE from SYSCONF: %zu Bytes\n", c_cl_size);
@@ -21,11 +21,11 @@ static inline uint64_t get_cl_size()
 
 #define CL_SIZE get_cl_size()
 
-uint64_t getsizeof(TestConfigStreamType);
-uint64_t getstreamelems(RuntimeStreamConfig *sdata);
-uint64_t getstreambytes(RuntimeStreamConfig *sdata);
+size_t getsizeof(TestConfigStreamType);
+size_t getstreamelems(RuntimeStreamConfig *sdata);
+size_t getstreambytes(RuntimeStreamConfig *sdata);
 
-int init_function(void* ptr, int state, TestConfigStreamType type, int dims, uint64_t* dimsizes, void* init_val, ...);
+int init_function(void* ptr, int state, TestConfigStreamType type, int dims, size_t* dimsizes, void* init_val, ...);
 
 int allocate_arrays(RuntimeStreamConfig *sdata);
 void release_arrays(RuntimeStreamConfig *sdata);
