+++ include/helper.h
@@ -3,6 +3,6 @@
 #define HELPER_H
 
 double convertToSeconds(const_bstring input);
-uint64_t convertToBytes(const_bstring input);
+size_t convertToBytes(const_bstring input);
 
 #endif /* HELPER_H */
